requirements.txt
astroid==2.9.2
bumpversion==0.6.0
check-manifest==0.47
coverage==6.2
docutils==0.18.1
factory-boy==3.2.1
Faker==11.3.0
flit==3.6.0
Pygments==2.18.0
pylint==2.12.2
pylint-django==2.5.0
python-dateutil==2.8.2
tox==3.24.5
